+++ html/layouts/chromes/default.php
@@ -19,8 +19,8 @@ if ($module->content === null || $module->content === '')
 
 $modulePos   = $module->position;
 $moduleTag   = $params->get('module_tag', 'div');
-$headerTag   = htmlspecialchars($params->get('header_tag', 'h4'), ENT_QUOTES, 'UTF-8');
-$headerClass = htmlspecialchars($params->get('header_class', ''), ENT_QUOTES, 'UTF-8');
+$headerTag   = htmlspecialchars($params->get('header_tag', 'h3'), ENT_QUOTES, 'UTF-8');
+$headerClass = htmlspecialchars($params->get('header_class', 'is-size-4'), ENT_QUOTES, 'UTF-8');
 
 ?>
 <<?php echo $moduleTag; ?> class="<?php echo $modulePos; ?> card <?php echo htmlspecialchars($params->get('moduleclass_sfx'), ENT_QUOTES, 'UTF-8'); ?>">