+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## WIP
 - Fix default "theme-switcher" parameter value
+- Fix default heading for modules
 
 ## 0.0.13
 - Add support for automatic intro/full text image resizing (Thanks @nikosdion)